requirements.txt
requests~=2.32.3
inquirer~=3.4.0
python-dotenv~=1.0.1
homeassistant-api~=5.0.0
